Add Interval

+++ BengalApplication/app/controllers/locations_controller.rb
@@ -7,12 +7,13 @@ class LocationsController < ApplicationController
   def new
     @occasion = Occasion.find(params[:occasion_id])
     @location = @occasion.locations.build
-    @location.time_slots.build
+    authorize @location
   end
 
   def create
     @occasion = Occasion.find(params[:occasion_id])
     @location = @occasion.locations.build(location_params)
+    authorize @location
     if @location.save
       redirect_to occasion_path(@occasion)
     else
@@ -24,11 +25,13 @@ class LocationsController < ApplicationController
   def show
     @occasion = Occasion.find_by(params[:occasion_id])
     @location = Location.find_by(params[:id])
+    authorize @location
   end
 
   def destroy
     @occasion = Occasion.find(params[:occasion_id])
     location = Location.find(params[:id])
+    authorize location
     location.destroy
     redirect_to occasion_path(@occasion)
   end
@@ -36,6 +39,7 @@ class LocationsController < ApplicationController
   private
 
   def location_params
-    params.require(:location).permit(:name, time_slots_attributes: [:id, :interval, :start_time, :end_time])
+    params.require(:location).permit(:name)
+      # params.require(:location).permit(:name, time_slots_attributes: [:id, :interval, :start_time, :end_time])
   end
 end
